Refs #106. fixing some test cases
